[EasyWebhook] Prevent extra results to be stored
