+++ packages/EasyWebhook/src/Middleware/AsyncMiddleware.php
@@ -10,6 +10,7 @@ use EonX\EasyWebhook\Interfaces\StackInterface;
 use EonX\EasyWebhook\Interfaces\Stores\StoreInterface;
 use EonX\EasyWebhook\Interfaces\WebhookInterface;
 use EonX\EasyWebhook\Interfaces\WebhookResultInterface;
+use EonX\EasyWebhook\ShouldNotBeStoredWebhookResult;
 use EonX\EasyWebhook\WebhookResult;
 
 final class AsyncMiddleware extends AbstractMiddleware
@@ -62,6 +63,6 @@ final class AsyncMiddleware extends AbstractMiddleware
 
         $this->dispatcher->dispatch($webhook);
 
-        return new WebhookResult($webhook);
+        return new ShouldNotBeStoredWebhookResult($webhook);
     }
 }
