+++ packages/EasyWebhook/src/ShouldNotBeStoredWebhookResult.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace EonX\EasyWebhook;
+
+final class ShouldNotBeStoredWebhookResult extends AbstractWebhookResult
+{
+    // No body needed.
+}
