+++ packages/EasyWebhook/tests/Middleware/StoreMiddlewareTest.php
@@ -4,7 +4,9 @@ declare(strict_types=1);
 
 namespace EonX\EasyWebhook\Tests\Middleware;
 
+use EonX\EasyWebhook\Interfaces\WebhookResultInterface;
 use EonX\EasyWebhook\Middleware\StoreMiddleware;
+use EonX\EasyWebhook\ShouldNotBeStoredWebhookResult;
 use EonX\EasyWebhook\Stores\ArrayResultStore;
 use EonX\EasyWebhook\Stores\ArrayStore;
 use EonX\EasyWebhook\Tests\AbstractMiddlewareTestCase;
@@ -12,22 +14,41 @@ use EonX\EasyWebhook\Webhook;
 
 final class StoreMiddlewareTest extends AbstractMiddlewareTestCase
 {
-    public function testProcess(): void
+    /**
+     * @return iterable<mixed>
+     */
+    public function providerTestProcess(): iterable
+    {
+        yield 'Should store' => [1];
+
+        yield 'Should not store' => [
+            0,
+            new ShouldNotBeStoredWebhookResult(Webhook::create('https://eonx.com')),
+        ];
+    }
+
+    /**
+     * @dataProvider providerTestProcess
+     */
+    public function testProcess(int $resultsCount, ?WebhookResultInterface $webhookResult = null): void
     {
         $webhook = new Webhook();
         $store = new ArrayStore($this->getRandomGenerator(), $this->getDataCleaner());
         $resultStore = new ArrayResultStore($this->getRandomGenerator(), $this->getDataCleaner());
         $middleware = new StoreMiddleware($store, $resultStore);
 
-        $result = $this->process($middleware, $webhook);
+        $result = $this->process($middleware, $webhook, $webhookResult);
         $webhooks = $store->getWebhooks();
         $results = $resultStore->getResults();
         $firstWebhook = \reset($webhooks);
         $firstResult = \reset($results);
 
         self::assertCount(1, $webhooks);
-        self::assertCount(1, $results);
         self::assertSame($firstWebhook, $webhook);
-        self::assertSame($firstResult, $result);
+        self::assertCount($resultsCount, $results);
+
+        if ($resultsCount > 0) {
+            self::assertSame($firstResult, $result);
+        }
     }
 }
